requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-admin-interface==0.29.2
django-cleanup==9.0.0
django-colorfield==0.11.0
django-grappelli==4.0.1
django-jazzmin==3.0.1
django-jet==1.0.8
django-modeltranslation==0.19.11
django-rosetta==0.10.1
django-ckeditor==6.7.1           
django-js-asset==3.0.1           
psycopg2-binary==2.9.9           
idna==3.10
pillow==11.0.0
polib==1.2.0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.2
text-unidecode==1.3
Unidecode==1.3.8
urllib3==2.2.3
psycopg2